% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2apath.R
\name{h2aPath}
\alias{h2aPath}
\title{Histone H2A data}
\usage{
h2aPath()
}
\value{
\code{character}, path to the directory containing the example
files.
}
\description{
This function returns the path to the Histone H2A example files accompanying
the \code{\link[topdown]{topdown-package}}.
}
\examples{
library("topdowndata")
h2aPath()
}
\seealso{
\code{\link[topdown]{topdown-package}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
