% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topDownDataPath.R
\name{topDownDataPath}
\alias{topDownDataPath}
\title{TopDown Proteomic Datasets}
\usage{
topDownDataPath(protein = c("myoglobin", "ca", "h3_3", "h4", "c3a"))
}
\arguments{
\item{protein}{\code{character}, name of the dataset.}
}
\value{
\code{character}, path to the directory containing the example
files.
}
\description{
This function returns the path to the example files accompanying
the \code{topdownr}.
}
\details{
See \link{topdownrdata-package} for a description of the datasets.
}
\examples{
topDownDataPath("myoglobin")
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
