% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2apath.R
\name{topDownDataPath}
\alias{topDownDataPath}
\title{TopDown Proteomic datasets}
\usage{
topDownDataPath(protein = c("myoglobin", "ca", "h3_3", "h4", "c3a"))
}
\arguments{
\item{protein}{\code{character}, name of the dataset.}
}
\value{
\code{character}, path to the directory containing the example
files.
}
\description{
This function returns the path to the example files accompanying
the \link[topdown:topdown-package]{topdown::topdown-package}.
}
\examples{
library("topdowndata")
topDownDataPath("myoglobin")
}
\seealso{
\link[topdown:topdown-package]{topdown::topdown-package}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
